Measurement development
Using random number generators to create fake values reported by the sensors (since I still don't have them).
The Measurement class knows what type of measurement it is; the user or the abstract sensor doesn't need to know what type of measurement it is as long as it can print its value (and know the type of the value).
Next, implement methods to get values every interval n seconds for duration x seconds.