Autosave on desync, etc.
Now the instructor will save files generated by students automatically when they disconnect.
Student GUIs will attempt to connect before constructing a screen. If an IOException is thrown by the socket constructor in ClientProxy, another exception is thrown to the GUI to indicate that a connection could not be made, and the program stops.